Tidy fixes

+++ soccer/planning/planner/EscapeObstaclesPathPlanner.hpp
@@ -54,7 +54,7 @@ public:
 
     static void createConfiguration(Configuration* cfg);
 
-    static float stepSize() { return *_stepSize; }
+    static double stepSize() { return *_stepSize; }
 
 private:
     PathTargetPlanner _planner;
